@@ -1,12 +1,12 @@
-const path = require('path')
-const cp = require('child_process')
+import path from "path"
+import cp from "child_process"
 
 const __path_comrak = path.join(__dirname, "./wasm/comrak")
 
-const __command_build = 'wasm-pack build --target web'
+const __command_build = "wasm-pack build --target web"
 
-cp.exec(__command_build, { cwd: __path_comrak, shell: 'powershell' }, function (err, stdout, stderr) {
+cp.exec(__command_build, { cwd: __path_comrak, shell: "powershell" }, function (err, stdout, stderr) {
     console.log(err)
     console.log(stdout)
     console.log(stderr)
-});
+})
