@@ -1,5 +1,5 @@
-const express = require('express')
+import express from "express"
 
 const app = express()
-app.use('/desq', express.static(`./dist/`))
+app.use("/desq", express.static(`./dist/`))
 app.listen(8080)
