@@ -1,11 +1,5 @@
 import React, { FC } from "react"
 
 export const Editor: FC = () => {
-
-    return (
-        <textarea className="h-fit w-fit" placeholder="Write your thoughts here...">
-
-        </textarea>
-    )
-
+    return <textarea className="h-fit w-fit" placeholder="Write your thoughts here..."></textarea>
 }
