@@ -1,28 +1,28 @@
-import { SideNavigation } from "./SideNavigation";
-import { useNavigate, useLocation } from "react-router-dom";
-import React, { FC, useState } from "react";
-import { SomeIcon } from "../Icon";
+import { SideNavigation } from "./SideNavigation"
+import React, { FC, useState } from "react"
+import { SomeIcon } from "../Icon"
+import { useLocation, useNavigate } from "../../../node_modules/react-router-dom/dist/index"
 
 export const NavSidebar: FC = () => {
-    const history = useNavigate();
-    const location = useLocation();
-    const [isSidebarOpen, setIsSidebarOpen] = useState(true);
+    const history = useNavigate()
+    const location = useLocation()
+    const [isSidebarOpen, setIsSidebarOpen] = useState(true)
 
     return (
         <>
             <div
-                className={`fixed inset-y-0 left-0 z-30 w-64 overflow-y-auto transition duration-300 ease-out transform translate-x-0 bg-white border-r-2 lg:translate-x-0 lg:static lg:inset-0 ${isSidebarOpen ? "ease-out translate-x-0" : "ease-in -translate-x-full"}`}
+                className={`fixed inset-y-0 left-0 z-30 w-64 overflow-y-auto transition duration-300 ease-out transform translate-x-0 bg-white border-r-2 lg:translate-x-0 lg:static lg:inset-0 ${
+                    isSidebarOpen ? "ease-out translate-x-0" : "ease-in -translate-x-full"
+                }`}
             >
                 <div className="flex items-center justify-center mt-10 text-center py-6">
-                    <span className="mx-2 text-2xl font-semibold text-black">
-                        Desq Web App
-                    </span>
+                    <span className="mx-2 text-2xl font-semibold text-black">Desq Web App</span>
                 </div>
 
                 <SideNavigation
                     activeItemId={location.pathname}
                     onSelect={({ itemId }) => {
-                        history(itemId);
+                        history(itemId)
                     }}
                     items={[
                         { title: "Home", itemId: "/" },
@@ -30,22 +30,20 @@ export const NavSidebar: FC = () => {
                         { title: "Fragments", itemId: "/fragments" },
                         { title: "Portraits", itemId: "/portraits" },
                         { title: "Podcasts", itemId: "/pods" },
-                        { title: "Images", itemId: "/new-image" }
+                        { title: "Images", itemId: "/new-image" },
                     ]}
                 />
 
                 <div className="absolute bottom-0 w-full my-8">
                     <SideNavigation
                         activeItemId={location.pathname}
-                        items={[
-                            { title: "Storage", itemId: "/storage", elemBefore: () => <SomeIcon /> }
-                        ]}
+                        items={[{ title: "Storage", itemId: "/storage", elemBefore: () => <SomeIcon /> }]}
                         onSelect={({ itemId }) => {
-                            history(itemId);
+                            history(itemId)
                         }}
                     />
                 </div>
             </div>
         </>
-    );
-};
+    )
+}
