@@ -6,7 +6,6 @@ import { AddressBar } from "./AddressBar"
 import { SaveIcon, UpRightArrow } from "../Icon"
 
 export const PageContainer: FC = () => {
-
     const size = useWindowSize()
     const ifrRef = useRef<HTMLIFrameElement>(null)
 
@@ -17,11 +16,13 @@ export const PageContainer: FC = () => {
                 const doc = ifr.contentWindow?.document
                 if (doc) {
                     doc.open()
-                    findOneFile('Hidden-subgroup-problem.mhtml')
-                        .then(blob => blob.text())
+                    findOneFile("Hidden-subgroup-problem.mhtml")
+                        .then((blob) => blob.text())
                         .then(mhtml2html.convert)
-                        .then(x => x?.window.document.documentElement.outerHTML)
-                        .then(x => { doc.write(x!) })
+                        .then((x) => x?.window.document.documentElement.outerHTML)
+                        .then((x) => {
+                            doc.write(x!)
+                        })
                         .then(() => doc.close())
                 }
             }
@@ -33,14 +34,20 @@ export const PageContainer: FC = () => {
         <div className="overflow-hidden shadow-2xl rounded-2xl " style={{ width: 1280, height: size.height * 0.9 }}>
             <div className="flex items-center pl-3 justify-between bg-gray-200 rounded-t-xl h-12">
                 <div className="flex h-[90%] align-middle items-center">
-                    {[1, 2, 3].map(_ => <span className="w-3 h-3 p-1 m-1 bg-white rounded-full"></span>)}
-                    <AddressBar placeHolder="url" onChange={() => { }} onSearch={e => {
-                        let url = e.target[0].value
-                        const ifr = ifrRef.current
-                        if (ifr) {
-                            ifr.src = url
-                        }
-                    }} />
+                    {[1, 2, 3].map((_) => (
+                        <span className="w-3 h-3 p-1 m-1 bg-white rounded-full"></span>
+                    ))}
+                    <AddressBar
+                        placeHolder="url"
+                        onChange={() => {}}
+                        onSearch={(e) => {
+                            let url = e.target[0].value
+                            const ifr = ifrRef.current
+                            if (ifr) {
+                                ifr.src = url
+                            }
+                        }}
+                    />
                 </div>
                 <div className="flex h-[90%] align-middle">
                     <button className="align-middle">
@@ -51,7 +58,7 @@ export const PageContainer: FC = () => {
                     </button>
                 </div>
             </div>
-            <iframe className="rounded-b-xl align-middle h-[calc(100vh_-_10.75rem)]" ref={ifrRef} style={{ width: '100%', height: '100%' }} />
+            <iframe className="rounded-b-xl align-middle h-[calc(100vh_-_10.75rem)]" ref={ifrRef} style={{ width: "100%", height: "100%" }} />
         </div>
     )
 }
