@@ -1,4 +1,4 @@
-import React from "react";
-import { AppRoutes } from "./_routes";
+import React from "react"
+import { AppRoutes } from "./_routes"
 
 export const _app = <AppRoutes />
