@@ -1,12 +1,13 @@
-import React, { FC } from "react";
-import { Routes, Route, HashRouter } from "react-router-dom";
-import { NotesPage } from "./notes";
-import { FragmentsPage } from "./fragments";
-import { PortraitsPage } from "./portraits";
-import { StoragePage } from "./storage";
-import { PodsPage } from "./pods";
-import { WelcomePage } from "./welcome";
-import { ImagePage } from "./image";
+import React, { FC } from "react"
+import { NotesPage } from "./notes"
+import { FragmentsPage } from "./fragments"
+import { PortraitsPage } from "./portraits"
+import { StoragePage } from "./storage"
+import { PodsPage } from "./pods"
+import { WelcomePage } from "./welcome"
+import { ImagePage } from "./image"
+import { HashRouter } from "../../node_modules/react-router-dom/dist/index"
+import { Routes, Route } from "../../node_modules/react-router/dist/index"
 
 export const AppRoutes: FC = () => {
     return (
@@ -21,6 +22,5 @@ export const AppRoutes: FC = () => {
                 <Route path="new-image" element={<ImagePage />} />
             </Routes>
         </HashRouter>
-    );
-};
-
+    )
+}
