@@ -1,8 +1,3 @@
-export const AZURE_SPEECH_API = 'https://eastasia.api.cognitive.microsoft.com/sts/v1.0/issuetoken'
-export const CORS_ANYWHERE = 'http://192.168.1.163:8881/'
-export const SITES = [
-    '.wikipedia.org',
-    '.zhihu.com',
-    '.reddit.com',
-    'other'
-] as const
+export const AZURE_SPEECH_API = "https://eastasia.api.cognitive.microsoft.com/sts/v1.0/issuetoken"
+export const CORS_ANYWHERE = "http://192.168.1.163:8881/"
+export const SITES = [".wikipedia.org", ".zhihu.com", ".reddit.com", "other"] as const
