@@ -1,14 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-	content: [
-		"./src/**/*.tsx",
-		"./resources/main.html"
-	],
-	theme: {
-		extend: {},
-	},
-	plugins: [
-		require('@tailwindcss/typography')
-	]
+    content: ["./src/**/*.tsx", "./resources/main.html"],
+    theme: {
+        extend: {},
+    },
+    plugins: [require("@tailwindcss/typography")],
 }
-
